requirements.txt
pytorch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
Pillow==9.5.0
git+https://github.com/openai/CLIP.git

